+++ users/client.js
@@ -1,0 +1,6 @@
+/**
+ * Check if the logged in users has permission for something.
+ */
+Template.registerHelper('doIHavePermission', function(key) {
+	return Meteor.user().hasPermission(key);
+});