+++ CHANGELOG.md
@@ -17,6 +17,10 @@ This is the changelog for ```orionjs:core```.
 
 - Reorder orion config code.
 
+- Reorder users code.
+
+- Added the posibility to modify the user object on creation using ```orion.users.onCreateUser(options, user)```. Thanks @splendido.
+
 ### 0.4.10
 
 - Fixed bug #21.